refactor to remove duplicated code
